@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.8.1",
+  "version": "8.10.0",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -17,47 +17,47 @@
     "orm",
     "db"
   ],
+  "type": "commonjs",
   "license": "MIT",
   "dependencies": {
-    "bson": "^6.7.0",
+    "bson": "^6.10.1",
     "kareem": "2.6.3",
-    "mongodb": "~6.10.0",
+    "mongodb": "~6.13.0",
     "mpath": "0.9.0",
     "mquery": "5.0.0",
     "ms": "2.1.3",
     "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.26.0",
-    "@babel/preset-env": "7.26.0",
-    "@typescript-eslint/eslint-plugin": "^8.4.0",
-    "@typescript-eslint/parser": "^8.4.0",
+    "@babel/core": "7.26.7",
+    "@babel/preset-env": "7.26.7",
+    "@typescript-eslint/eslint-plugin": "^8.19.1",
+    "@typescript-eslint/parser": "^8.19.1",
     "acquit": "1.3.0",
     "acquit-ignore": "0.2.1",
     "acquit-require": "0.1.1",
+    "ajv": "8.17.1",
     "assert-browserify": "2.0.0",
-    "axios": "1.1.3",
     "babel-loader": "8.2.5",
     "broken-link-checker": "^0.7.8",
     "buffer": "^5.6.0",
     "cheerio": "1.0.0",
     "crypto-browserify": "3.12.1",
-    "dotenv": "16.4.5",
     "dox": "1.0.0",
-    "eslint": "8.57.0",
-    "eslint-plugin-markdown": "^5.0.0",
+    "eslint": "8.57.1",
+    "eslint-plugin-markdown": "^5.1.0",
     "eslint-plugin-mocha-no-only": "1.2.0",
     "express": "^4.19.2",
-    "fs-extra": "~11.2.0",
-    "highlight.js": "11.10.0",
+    "fs-extra": "~11.3.0",
+    "highlight.js": "11.11.1",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
-    "markdownlint-cli2": "^0.14.0",
-    "marked": "14.1.3",
+    "markdownlint-cli2": "^0.17.1",
+    "marked": "15.0.6",
     "mkdirp": "^3.0.1",
-    "mocha": "10.8.2",
+    "mocha": "11.1.0",
     "moment": "2.30.1",
-    "mongodb-memory-server": "10.1.2",
+    "mongodb-memory-server": "10.1.3",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.3",
@@ -65,9 +65,9 @@
     "sinon": "19.0.2",
     "stream-browserify": "3.0.0",
     "tsd": "0.31.2",
-    "typescript": "5.6.3",
-    "uuid": "11.0.2",
-    "webpack": "5.96.1"
+    "typescript": "5.7.3",
+    "uuid": "11.0.5",
+    "webpack": "5.97.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
@@ -104,6 +104,8 @@
     "test-deno": "deno run --allow-env --allow-read --allow-net --allow-run --allow-sys --allow-write ./test/deno.js",
     "test-rs": "START_REPLICA_SET=1 mocha --timeout 30000 --exit ./test/*.test.js",
     "test-tsd": "node ./test/types/check-types-filename && tsd",
+    "setup-test-encryption": "bash scripts/configure-cluster-with-encryption.sh",
+    "test-encryption": "mocha --exit ./test/encryption/*.test.js",
     "tdd": "mocha ./test/*.test.js --inspect --watch --recursive --watch-files ./**/*.{js,ts}",
     "test-coverage": "nyc --reporter=html --reporter=text npm test",
     "ts-benchmark": "cd ./benchmarks/typescript/simple && npm install && npm run benchmark | node ../../../scripts/tsc-diagnostics-check"
