@@ -1,8 +1,8 @@
 {
   "name": "tr46",
-  "version": "4.1.1",
+  "version": "5.0.0",
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   },
   "description": "An implementation of the Unicode UTS #46: Unicode IDNA Compatibility Processing",
   "main": "index.js",
@@ -11,7 +11,7 @@
     "lib/"
   ],
   "scripts": {
-    "test": "mocha",
+    "test": "node --test",
     "lint": "eslint .",
     "pretest": "node scripts/getLatestTests.js",
     "prepublish": "node scripts/generateMappingTable.js && node scripts/generateRegexes.js"
@@ -31,15 +31,13 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "punycode": "^2.3.0"
+    "punycode": "^2.3.1"
   },
   "devDependencies": {
     "@domenic/eslint-config": "^3.0.0",
-    "@unicode/unicode-15.0.0": "^1.3.1",
-    "eslint": "^8.32.0",
-    "minipass-fetch": "^3.0.1",
-    "mocha": "^10.2.0",
+    "@unicode/unicode-15.1.0": "^1.5.2",
+    "eslint": "^8.53.0",
     "regenerate": "^1.4.2"
   },
-  "unicodeVersion": "15.0.0"
+  "unicodeVersion": "15.1.0"
 }
