@@ -1,6 +1,6 @@
 {
   "name": "whatwg-url",
-  "version": "13.0.0",
+  "version": "14.1.1",
   "description": "An implementation of the WHATWG URL Standard's URL API and parsing machinery",
   "main": "index.js",
   "files": [
@@ -12,47 +12,42 @@
   "license": "MIT",
   "repository": "jsdom/whatwg-url",
   "dependencies": {
-    "tr46": "^4.1.1",
+    "tr46": "^5.0.0",
     "webidl-conversions": "^7.0.0"
   },
   "devDependencies": {
-    "@domenic/eslint-config": "^3.0.0",
+    "@domenic/eslint-config": "^4.0.1",
     "benchmark": "^2.1.4",
-    "domexception": "^4.0.0",
-    "esbuild": "^0.17.19",
-    "eslint": "^8.41.0",
-    "jest": "^29.5.0",
-    "minipass-fetch": "^3.0.3",
-    "webidl2js": "^17.1.0"
+    "c8": "^10.1.2",
+    "esbuild": "^0.24.0",
+    "eslint": "^9.16.0",
+    "globals": "^15.13.0",
+    "webidl2js": "^18.0.0"
   },
   "engines": {
-    "node": ">=16"
+    "node": ">=18"
   },
   "scripts": {
-    "coverage": "jest --coverage",
-    "lint": "eslint .",
+    "coverage": "c8 node --test --experimental-test-coverage test/*.js",
+    "lint": "eslint",
     "prepare": "node scripts/transform.js",
     "pretest": "node scripts/get-latest-platform-tests.js && node scripts/transform.js",
     "build-live-viewer": "esbuild --bundle --format=esm --sourcemap --outfile=live-viewer/whatwg-url.mjs index.js",
-    "test": "jest"
+    "test": "node --test test/*.js",
+    "bench": "node scripts/benchmark.js"
   },
-  "jest": {
-    "collectCoverageFrom": [
-      "lib/**/*.js",
-      "!lib/utils.js"
+  "c8": {
+    "reporter": [
+      "text",
+      "html"
     ],
-    "coverageDirectory": "coverage",
-    "coverageReporters": [
-      "lcov",
-      "text-summary"
-    ],
-    "testEnvironment": "node",
-    "testMatch": [
-      "<rootDir>/test/**/*.js"
-    ],
-    "testPathIgnorePatterns": [
-      "^<rootDir>/test/testharness.js$",
-      "^<rootDir>/test/web-platform-tests/"
+    "exclude": [
+      "lib/Function.js",
+      "lib/URL.js",
+      "lib/URLSearchParams.js",
+      "lib/utils.js",
+      "scripts/",
+      "test/"
     ]
   }
 }
