@@ -1,61 +1,23 @@
-// const express = require('express');
-// import dotenv from "dotenv";
-// import cookieParser from "cookie-parser";
-// import cors from "cors";
+import express from 'express';
+import dotenv from 'dotenv';
+import cookieParser from 'cookie-parser';
+import cors from 'cors';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import { connectDB } from './lib/db.js'; // Adjust the path as needed
+import authRoutes from './routes/auth.route.js'; // Adjust the path as needed
+import messageRoutes from './routes/message.route.js'; // Adjust the path as needed
+import { app, server } from './lib/socket.js'; // Adjust the path as needed
 
-// import path from "path";
-
-// import { connectDB } from "./lib/db.js";
-
-// import authRoutes from "./routes/auth.route.js";
-// import messageRoutes from "./routes/message.route.js";
-// import { app, server } from "./lib/socket.js";
-
-// dotenv.config();
-
-// const PORT = process.env.PORT;
-// const __dirname = path.resolve();
-
-// app.use(express.json());
-// app.use(cookieParser());
-// app.use(
-//   cors({
-//     origin: ["http://localhost:5173", "http://localhost:3000", "https://spawn-nine.vercel.app"],
-//     credentials: true,
-//   })
-// );
-
-
-// app.use("/api/auth", authRoutes);
-// app.use("/api/messages", messageRoutes);
-
-// if (process.env.NODE_ENV === "production") {
-//   app.use(express.static(path.join(__dirname, "../frontend/dist")));
-
-//   app.get("*", (req, res) => {
-//     res.sendFile(path.join(__dirname, "../frontend", "dist", "index.html"));
-//   });
-// }
-
-// server.listen(PORT, () => {
-//   console.log("server is running on PORT:" + PORT);
-//   connectDB();
-// });
-const express = require('express');
-const dotenv = require('dotenv');
-const cookieParser = require('cookie-parser');
-const cors = require('cors');
-const path = require('path');
-const { connectDB } = require('./lib/db');
-const authRoutes = require('./routes/auth.route');
-const messageRoutes = require('./routes/message.route');
-const { app, server } = require('./lib/socket');
-
+// Configure environment variables
 dotenv.config();
 
-const PORT = process.env.PORT;
-const __dirname = path.resolve();
+// Fix __dirname for ES modules
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
+// Middleware
 app.use(express.json());
 app.use(cookieParser());
 app.use(
@@ -65,9 +27,11 @@ app.use(
   })
 );
 
+// Routes
 app.use('/api/auth', authRoutes);
 app.use('/api/messages', messageRoutes);
 
+// Serve static files in production
 if (process.env.NODE_ENV === 'production') {
   app.use(express.static(path.join(__dirname, '../frontend/dist')));
 
@@ -76,5 +40,12 @@ if (process.env.NODE_ENV === 'production') {
   });
 }
 
+// Start the server
+const PORT = process.env.PORT || 5000;
+server.listen(PORT, () => {
+  console.log(`Server is running on port ${PORT}`);
+  connectDB()
+});
+
 // Export the app for Vercel
-module.exports = app;
+export default app;
