@@ -1,6 +1,6 @@
 import { Server } from "socket.io";
 import http from "http";
-const express = require('express');
+import express from 'express';
 
 const app = express();
 const server = http.createServer(app);
