@@ -1,4 +1,4 @@
-const express = require('express');
+import express from 'express';
 import { checkAuth, login, logout, signup, updateProfile } from "../controllers/auth.controller.js";
 import { protectRoute } from "../middleware/auth.middleware.js";
 
