@@ -1,4 +1,4 @@
-const express = require('express');
+import express from 'express';
 import { protectRoute } from "../middleware/auth.middleware.js";
 import { getMessages, getUsersForSidebar, sendMessage } from "../controllers/message.controller.js";
 
